Chat lists (#28)
* other: added a helper function for notifications

It returns a publisher for a given notification. Why create such function? NotificationCenter.default.publisher(for:) is too long, so I replace it with nsPublisher(for:)

* feat: started work on new chat list handler

* xcode: bumped project version

* fix: fixed the main chat list
Now it really shows the main chat list instead of every chat that this app had received
And also made a foundation for implementing other chat lists

* feat: added a button to show archive chat list

* feat: moved updateNewMessage receiver to ChatView

* fix: fix chat view input field border, now it won't lead perfectionists to a suicide

* fix: fix tiny chat list bug

* feat: improved TDLib encryption key generation using CryptoKit

* feat: added Generated module
This module will contain everything related to GYB generated files, like secrets, and a newly created String extension.

* feat: added config

* feat: finished code generation

* feat: small UI improvement of the archive chat list toggle